+++ app/utils/chart-config/versioning.ts
@@ -1534,6 +1534,29 @@ export const chartConfigMigrations: Migration[] = [
       return newConfig;
     },
   },
+  {
+    from: "4.4.0",
+    to: "4.5.0",
+    description: `all {
+      + conversionUnitsByComponentId
+    }`,
+    up: (config) => {
+      const newConfig = { ...config, version: "4.5.0" };
+
+      newConfig.conversionUnitsByComponentId = {};
+
+      return newConfig;
+    },
+    down: (config) => {
+      const newConfig = { ...config, version: "4.4.0" };
+
+      if (newConfig.conversionUnitsByComponentId) {
+        delete newConfig.conversionUnitsByComponentId;
+      }
+
+      return newConfig;
+    },
+  },
 ];
 
 export const migrateChartConfig = makeMigrate<ChartConfig>(
@@ -2146,6 +2169,12 @@ export const configuratorStateMigrations: Migration[] = [
     toVersion: "4.5.0",
     fromChartConfigVersion: "4.3.0",
     toChartConfigVersion: "4.4.0",
+  }),
+  makeBumpChartConfigVersionMigration({
+    fromVersion: "4.5.0",
+    toVersion: "4.6.0",
+    fromChartConfigVersion: "4.4.0",
+    toChartConfigVersion: "4.5.0",
   }),
 ];